requirements.txt
# Create conda env with python=3.10
matplotlib==3.9.0
numpy==1.26.4
pandas==2.2.2
pyarrow==16.0.0
scikit-learn==1.5.2
tqdm==4.64.1
plotly==5.22.0
neurokit2==0.2.10
tsfresh==0.20.2
polars[all]==1.26.0
ipywidgets==8.1.3
# sktime==0.33.1
ipykernel==6.29.5
hydra-core
scipy==1.14.0
pytorch-tcn==1.2.1
torch==2.5.1
tsai==0.4.0
joblib==1.4.2
pypg
torch
